use BigInteger as ID, and clean up unused migration sql

+++ backend/app/models/record.py
@@ -5,11 +5,11 @@ class Record(db.Model):
     """记录数据模型"""
     __tablename__ = 'records'
     
-    id = db.Column(db.Integer, primary_key=True)
+    id = db.Column(db.BigInteger, primary_key=True)
     content = db.Column(db.Text, nullable=False)
     category = db.Column(db.String(20), default='general')  # idea/task/note/general
-    parent_id = db.Column(db.Integer, db.ForeignKey('records.id'), nullable=True)  # 父任务ID，支持子任务
-    user_id = db.Column(db.Integer, db.ForeignKey('users.id'), nullable=True)  # 用户ID外键
+    parent_id = db.Column(db.BigInteger, db.ForeignKey('records.id'), nullable=True)  # 父任务ID，支持子任务
+    user_id = db.Column(db.BigInteger, db.ForeignKey('users.id'), nullable=True)  # 用户ID外键
     priority = db.Column(db.String(20), default='medium')  # low/medium/high/urgent
     progress = db.Column(db.Integer, default=0)  # 进度百分比 0-100
     progress_notes = db.Column(db.Text, nullable=True)  # 进展记录和问题描述
